# Make startup 2.5x faster by only importing `verifiers` when needed