@willccbb 2025-08-13
hm do you already have the full extras set installed for verifiers in your venv? prime-rl only needs the minimal verifiers deps, which shouldn't introduce too much overhead (currently it's just openai/datasets/pydantic/jinja2/rich)