@vgel 2025-08-13
> hm do you already have the full extras set installed for verifiers in your venv? prime-rl only needs the minimal verifiers deps, which shouldn't introduce too much overhead (currently it's just openai/datasets/pydantic/jinja2/rich)
> 
> maybe there's a better way to do lazy imports within verifiers, not sure that just moving it further in is what we want to do, it ideally _shouldn't_ be slow to import if not using it for training

This was using the venv from a fresh clone + `uv sync`. Looking at the uv tree I didn't see torch or anything ridiculous like that, but I saw datasets, and that pulled in pandas and pyarrow etc.